+++ flutter_mediapipe_vision_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+* Added `detectOnPlanes`.
+
 ## 0.2.1
 
 * Pass the image with `createImageBitmap(blob)` instead of `HTMLImageElement`.